Added an abstraction level for ModelFactory.
